+++ src/components/Sidebar.tsx
@@ -879,6 +879,10 @@ export function Sidebar({ sidebarRef, toggleModal, currentThemeName, setCurrentT
 					<${SVG_MobileSidebar}/>
 				</button>
 			</div>
+			<button
+				onClick=${() => toggleModal("about")}>
+				About
+			</button>
 			${!!lastError && html`
 				<span className="error-text">${lastError}</span>`}
 		</div>